# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Dokumentasi API untuk aplikasi Laravel ini.",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://localhost:8000/api"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:800/api",
            "description": "Server API utama"
        }
    ],
    "paths": {
...
        "/api/expense/{expenseId}": {
            "get": {
                "tags": [
                    "Expense"
                ],
                "summary": "Get expense by ID",
                "description": "Mengambil detail expense berdasarkan ID, termasuk status dan approval stages.",
                "operationId": "31447720c1915a0d85cf2225641b9bfc",
                "parameters": [
                    {
                        "name": "expenseId",
                        "in": "path",
                        "description": "ID dari expense yang ingin diambil",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data expense",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 200
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "success"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Berhasil ambil data expense"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 9
                                                },
                                                "amount": {
                                                    "type": "string",
                                                    "format": "decimal",
                                                    "example": "10.00"
                                                },
                                                "status": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 2
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "disetujui"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "approval": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "approver": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer",
                                                                        "example": 3
                                                                    },
                                                                    "name": {
                                                                        "type": "string",
                                                                        "example": "xavier"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            },
                                                            "status": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer",
                                                                        "example": 2
                                                                    },
                                                                    "name": {
                                                                        "type": "string",
                                                                        "example": "disetujui"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Expense tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 404
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "error"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "No query results for model [App\\\\Models\\\\Expense] 9"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/expense": {
            "post": {
                "tags": [
                    "Expense"
                ],
                "summary": "Membuat expense baru",
                "operationId": "a8e3ae7ca07f1c973a526d413dc78a7e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1000
                                    },
                                    "status_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Expense berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 201
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "success"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Expense created successfully"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "data": {
                                            "properties": {
                                                "amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 10
                                                },
                                                "status_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "id": {
                                                    "type": "integer",
                                                    "example": 9
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 422
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "error"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Validation failed"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/expense/{expense}/approve": {
            "post": {
                "tags": [
                    "Expense"
                ],
                "summary": "Meng-approve expense",
                "operationId": "b616365897cd830a8bfc631183be363c",
                "parameters": [
                    {
                        "name": "expense",
                        "in": "path",
                        "description": "ID expense yang akan di-approve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "approver_id"
                                ],
                                "properties": {
                                    "approver_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Approval expense berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 200
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "success"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Approval expense berhasil"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "amount": {
                                                    "type": "decimal",
                                                    "example": 1
                                                },
                                                "status_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "approval": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "expense_id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "approver_id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "status_id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Expense tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        }
    },
...
}